feat: implement recursive reference rehydration and refactor typed engine
