+++ src/schema_service/server.rs
@@ -1,867 +1,22 @@
 use actix_cors::Cors;
-use actix_web::{web, App, HttpResponse, HttpServer as ActixHttpServer, Responder};
-use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
-use std::sync::{Arc, RwLock};
+use actix_web::{web, App, HttpServer as ActixHttpServer};
 
 use crate::error::{FoldDbError, FoldDbResult};
 use crate::log_feature;
 use crate::logging::features::LogFeature;
-use crate::schema::types::Schema;
-#[cfg(feature = "aws-backend")]
-use crate::storage::DynamoDbSchemaStore;
+
+// Re-export types for backward compatibility (e.g., schema_client.rs imports from here)
+pub use super::state::{SchemaServiceState, SchemaStorage};
+pub use super::types::*;
 
 #[cfg(feature = "aws-backend")]
-pub use crate::storage::CloudConfig;
-
-/// Response containing a list of available schema names
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SchemasListResponse {
-    pub schemas: Vec<String>,
-}
-
-/// Response containing all available schemas with their definitions
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AvailableSchemasResponse {
-    pub schemas: Vec<Schema>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SchemaSimilarityResponse {
-    pub similarity: f64,
-    pub closest_schema: Schema,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum SchemaAddOutcome {
-    Added(Schema, HashMap<String, String>), // Schema and mutation_mappers
-    AlreadyExists(Schema),                  // Exact same identity hash
-    TooSimilar(SchemaSimilarityResponse),
-}
-
-/// Error response structure
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ErrorResponse {
-    pub error: String,
-}
-
-/// Request structure for adding a schema with mutation mappers
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AddSchemaRequest {
-    pub schema: Schema,
-    pub mutation_mappers: HashMap<String, String>,
-}
-
-/// Response structure for adding a schema with mutation mappers
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AddSchemaResponse {
-    pub schema: Schema,
-    pub mutation_mappers: HashMap<String, String>,
-}
-
-/// Reload response structure
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ReloadResponse {
-    pub success: bool,
-    pub schemas_loaded: usize,
-}
-
-/// Health check response structure
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct HealthResponse {
-    pub status: String,
-}
-
-/// Conflict response for similar schemas
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ConflictResponse {
-    pub error: String,
-    pub similarity: f64,
-    pub closest_schema: Schema,
-}
-
-/// A schema entry with its similarity score
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SimilarSchemaEntry {
-    pub schema: Schema,
-    pub similarity: f64,
-}
-
-/// Response for the find-similar-schemas endpoint
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SimilarSchemasResponse {
-    pub query_schema: String,
-    pub threshold: f64,
-    pub similar_schemas: Vec<SimilarSchemaEntry>,
-}
-
-/// Storage backend for the schema service
-#[derive(Clone)]
-pub enum SchemaStorage {
-    /// Local sled database (default)
-    Sled {
-        db: sled::Db,
-        schemas_tree: sled::Tree,
-    },
-    /// Cloud storage (DynamoDB etc) (serverless, no locking needed!)
-    #[cfg(feature = "aws-backend")]
-    Cloud { store: Arc<DynamoDbSchemaStore> },
-}
-
-/// Shared state for the schema service
-#[derive(Clone)]
-pub struct SchemaServiceState {
-    schemas: Arc<RwLock<HashMap<String, Schema>>>,
-    storage: SchemaStorage,
-}
-
-impl SchemaServiceState {
-    /// Create a new schema service state with local sled storage
-    pub fn new(db_path: String) -> FoldDbResult<Self> {
-        let db = sled::open(&db_path).map_err(|e| {
-            FoldDbError::Config(format!(
-                "Failed to open schema service database at '{}': {}",
-                db_path, e
-            ))
-        })?;
-
-        let schemas_tree = db
-            .open_tree("schemas")
-            .map_err(|e| FoldDbError::Config(format!("Failed to open schemas tree: {}", e)))?;
-
-        let state = Self {
-            schemas: Arc::new(RwLock::new(HashMap::new())),
-            storage: SchemaStorage::Sled { db, schemas_tree },
-        };
-
-        // Load schemas synchronously for sled
-        state.load_schemas_sync()?;
-
-        Ok(state)
-    }
-
-    /// Synchronous version of load_schemas for Sled storage
-    fn load_schemas_sync(&self) -> FoldDbResult<()> {
-        let mut schemas = self
-            .schemas
-            .write()
-            .map_err(|_| FoldDbError::Config("Failed to acquire schemas write lock".to_string()))?;
-
-        schemas.clear();
-
-        match &self.storage {
-            SchemaStorage::Sled { schemas_tree, .. } => {
-                let mut count = 0;
-                for result in schemas_tree.iter() {
-                    let (key, value) = result.map_err(|e| {
-                        FoldDbError::Config(format!("Failed to iterate over schemas tree: {}", e))
-                    })?;
-
-                    let name = String::from_utf8(key.to_vec()).map_err(|e| {
-                        FoldDbError::Config(format!("Failed to decode schema name from key: {}", e))
-                    })?;
-
-                    let schema: Schema = serde_json::from_slice(&value).map_err(|e| {
-                        FoldDbError::Config(format!(
-                            "Failed to parse schema '{}' from database: {}",
-                            name, e
-                        ))
-                    })?;
-
-                    schemas.insert(name, schema);
-                    count += 1;
-                }
-
-                log_feature!(
-                    LogFeature::Schema,
-                    info,
-                    "Schema service loaded {} schemas from sled",
-                    count
-                );
-            }
-            #[cfg(feature = "aws-backend")]
-            _ => {
-                return Err(FoldDbError::Config(
-                    "load_schemas_sync called on non-Sled storage".to_string(),
-                ));
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Create a new schema service state with Cloud storage
-    /// No locking needed - identity hashes ensure idempotent writes!
-    #[cfg(feature = "aws-backend")]
-    pub async fn new_with_cloud(config: CloudConfig) -> FoldDbResult<Self> {
-        log_feature!(
-            LogFeature::Schema,
-            info,
-            "Initializing schema service with DynamoDB in region: {}",
-            config.region
-        );
-
-        let store = DynamoDbSchemaStore::new(config).await?;
-
-        let state = Self {
-            schemas: Arc::new(RwLock::new(HashMap::new())),
-            storage: SchemaStorage::Cloud {
-                store: Arc::new(store),
-            },
-        };
-
-        // Load schemas on initialization
-        state.load_schemas().await?;
-
-        log_feature!(
-            LogFeature::Schema,
-            info,
-            "Schema service initialized with DynamoDB, loaded {} schemas",
-            state.schemas.read().map(|s| s.len()).unwrap_or(0)
-        );
-
-        Ok(state)
-    }
-
-    /// Load all schemas from storage (works for both Sled and DynamoDB)
-    pub async fn load_schemas(&self) -> FoldDbResult<()> {
-        match &self.storage {
-            SchemaStorage::Sled { schemas_tree, .. } => {
-                let mut schemas = self.schemas.write().map_err(|_| {
-                    FoldDbError::Config("Failed to acquire schemas write lock".to_string())
-                })?;
-
-                schemas.clear();
-                let mut count = 0;
-                for result in schemas_tree.iter() {
-                    let (key, value) = result.map_err(|e| {
-                        FoldDbError::Config(format!("Failed to iterate over schemas tree: {}", e))
-                    })?;
-
-                    let name = String::from_utf8(key.to_vec()).map_err(|e| {
-                        FoldDbError::Config(format!("Failed to decode schema name from key: {}", e))
-                    })?;
-
-                    let schema: Schema = serde_json::from_slice(&value).map_err(|e| {
-                        FoldDbError::Config(format!(
-                            "Failed to parse schema '{}' from database: {}",
-                            name, e
-                        ))
-                    })?;
-
-                    log_feature!(
-                        LogFeature::Schema,
-                        info,
-                        "Loaded schema '{}' from sled database",
-                        name
-                    );
-
-                    schemas.insert(name, schema);
-                    count += 1;
-                }
-
-                log_feature!(
-                    LogFeature::Schema,
-                    info,
-                    "Schema service loaded {} schemas from sled",
-                    count
-                );
-            }
-            #[cfg(feature = "aws-backend")]
-            SchemaStorage::Cloud { store } => {
-                let all_schemas = store.get_all_schemas().await?;
-                let count = all_schemas.len();
-
-                let mut schemas = self.schemas.write().map_err(|_| {
-                    FoldDbError::Config("Failed to acquire schemas write lock".to_string())
-                })?;
-
-                schemas.clear();
-
-                for schema in all_schemas {
-                    log_feature!(
-                        LogFeature::Schema,
-                        info,
-                        "Loaded schema '{}' from DynamoDB",
-                        schema.name
-                    );
-                    schemas.insert(schema.name.clone(), schema);
-                }
-
-                log_feature!(
-                    LogFeature::Schema,
-                    info,
-                    "Schema service loaded {} schemas from DynamoDB",
-                    count
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    pub async fn add_schema(
-        &self,
-        mut schema: Schema,
-        mutation_mappers: HashMap<String, String>,
-    ) -> FoldDbResult<SchemaAddOutcome> {
-        // Ensure identity_hash is computed
-        if schema.identity_hash.is_none() {
-            schema.compute_identity_hash();
-        }
-
-        // Get the original schema name before we modify it
-        let original_schema_name = schema.name.clone();
-
-        // Use identity_hash as the schema identifier
-        let identity_hash = schema
-            .get_identity_hash()
-            .ok_or_else(|| {
-                FoldDbError::Config("Schema must have identity_hash computed".to_string())
-            })?
-            .clone();
-
-        log_feature!(
-            LogFeature::Schema,
-            info,
-            "Schema '{}' identity_hash: {}",
-            original_schema_name,
-            identity_hash
-        );
-
-        // Use identity_hash as unique identifier (SHA256 of sorted field names)
-        let schema_name = identity_hash.clone();
-
-        // Check if this exact schema already exists (same identity hash = same fields)
-        {
-            let schemas = self.schemas.read().map_err(|_| {
-                FoldDbError::Config("Failed to acquire schemas read lock".to_string())
-            })?;
-
-            if let Some(existing_schema) = schemas.get(&schema_name) {
-                log_feature!(
-                    LogFeature::Schema,
-                    info,
-                    "Schema '{}' already exists - returning existing schema",
-                    schema_name
-                );
-
-                return Ok(SchemaAddOutcome::AlreadyExists(existing_schema.clone()));
-            }
-        }
-
-        schema.name = schema_name.clone();
-
-        // Persist to storage backend
-        match &self.storage {
-            SchemaStorage::Sled { db, schemas_tree } => {
-                let serialized_schema = serde_json::to_vec(&schema).map_err(|error| {
-                    FoldDbError::Serialization(format!(
-                        "Failed to serialize schema '{}': {}",
-                        schema_name, error
-                    ))
-                })?;
-
-                schemas_tree
-                    .insert(schema_name.as_bytes(), serialized_schema)
-                    .map_err(|error| {
-                        FoldDbError::Config(format!(
-                            "Failed to insert schema '{}' into sled database: {}",
-                            schema_name, error
-                        ))
-                    })?;
-
-                db.flush().map_err(|error| {
-                    FoldDbError::Config(format!("Failed to flush sled database: {}", error))
-                })?;
-
-                log_feature!(
-                    LogFeature::Schema,
-                    info,
-                    "Schema '{}' persisted to sled database",
-                    schema_name
-                );
-            }
-            #[cfg(feature = "aws-backend")]
-            SchemaStorage::Cloud { store } => {
-                // No locking needed! Identity hash ensures idempotent writes
-                store.put_schema(&schema, &mutation_mappers).await?;
-
-                log_feature!(
-                    LogFeature::Schema,
-                    info,
-                    "Schema '{}' persisted to DynamoDB (no locking needed!)",
-                    schema_name
-                );
-            }
-        }
-
-        // Insert into in-memory cache
-        {
-            let mut schemas = self.schemas.write().map_err(|_| {
-                FoldDbError::Config("Failed to acquire schemas write lock".to_string())
-            })?;
-            schemas.insert(schema_name.clone(), schema.clone());
-        }
-
-        log_feature!(
-            LogFeature::Schema,
-            info,
-            "Schema '{}' successfully added to registry",
-            schema_name
-        );
-
-        Ok(SchemaAddOutcome::Added(schema, mutation_mappers))
-    }
-
-    /// Get all schema names (public accessor for Lambda integration)
-    pub fn get_schema_names(&self) -> FoldDbResult<Vec<String>> {
-        let schemas = self
-            .schemas
-            .read()
-            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
-        Ok(schemas.keys().cloned().collect())
-    }
-
-    /// Get all schemas (public accessor for Lambda integration)
-    pub fn get_all_schemas_cached(&self) -> FoldDbResult<Vec<Schema>> {
-        let schemas = self
-            .schemas
-            .read()
-            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
-        Ok(schemas.values().cloned().collect())
-    }
-
-    /// Get a schema by name (public accessor for Lambda integration)
-    pub fn get_schema_by_name(&self, name: &str) -> FoldDbResult<Option<Schema>> {
-        let schemas = self
-            .schemas
-            .read()
-            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
-        Ok(schemas.get(name).cloned())
-    }
-
-    /// Get schema count (public accessor for Lambda integration)
-    pub fn get_schema_count(&self) -> usize {
-        self.schemas.read().map(|s| s.len()).unwrap_or(0)
-    }
-
-    /// Find schemas similar to the given schema using Jaccard index on field name sets
-    pub fn find_similar_schemas(
-        &self,
-        name: &str,
-        threshold: f64,
-    ) -> FoldDbResult<SimilarSchemasResponse> {
-        let schemas = self
-            .schemas
-            .read()
-            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
-
-        let target = schemas.get(name).ok_or_else(|| {
-            FoldDbError::Config(format!("Schema '{}' not found", name))
-        })?;
-
-        let target_fields = collect_field_names(target);
-
-        let mut similar: Vec<SimilarSchemaEntry> = schemas
-            .iter()
-            .filter(|(k, _)| k.as_str() != name)
-            .filter_map(|(_, schema)| {
-                let other_fields = collect_field_names(schema);
-                let similarity = jaccard_index(&target_fields, &other_fields);
-                if similarity >= threshold {
-                    Some(SimilarSchemaEntry {
-                        schema: schema.clone(),
-                        similarity,
-                    })
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        similar.sort_by(|a, b| b.similarity.partial_cmp(&a.similarity).unwrap_or(std::cmp::Ordering::Equal));
-
-        Ok(SimilarSchemasResponse {
-            query_schema: name.to_string(),
-            threshold,
-            similar_schemas: similar,
-        })
-    }
-}
-
-/// Collect all field names from a schema (union of fields and transform_fields keys)
-fn collect_field_names(schema: &Schema) -> HashSet<String> {
-    let mut names = HashSet::new();
-    if let Some(ref fields) = schema.fields {
-        for f in fields {
-            names.insert(f.clone());
-        }
-    }
-    if let Some(ref tf) = schema.transform_fields {
-        for key in tf.keys() {
-            names.insert(key.clone());
-        }
-    }
-    names
-}
-
-/// Compute Jaccard index: |A ∩ B| / |A ∪ B|
-fn jaccard_index(a: &HashSet<String>, b: &HashSet<String>) -> f64 {
-    if a.is_empty() && b.is_empty() {
-        return 1.0;
-    }
-    let intersection = a.intersection(b).count();
-    let union = a.union(b).count();
-    intersection as f64 / union as f64
-}
-
-/// List all available schemas
-async fn list_schemas(state: web::Data<SchemaServiceState>) -> impl Responder {
-    let schemas = match state.schemas.read() {
-        Ok(s) => s,
-        Err(e) => {
-            log_feature!(
-                LogFeature::Schema,
-                error,
-                "Failed to acquire schemas read lock: {}",
-                e
-            );
-            return HttpResponse::InternalServerError().json(ErrorResponse {
-                error: "Failed to acquire schemas read lock".to_string(),
-            });
-        }
-    };
-
-    let schema_names: Vec<String> = schemas.keys().cloned().collect();
-
-    HttpResponse::Ok().json(SchemasListResponse {
-        schemas: schema_names,
-    })
-}
-
-/// Get all available schemas with their full definitions
-async fn get_available_schemas(state: web::Data<SchemaServiceState>) -> impl Responder {
-    let schemas = match state.schemas.read() {
-        Ok(s) => s,
-        Err(e) => {
-            log_feature!(
-                LogFeature::Schema,
-                error,
-                "Failed to acquire schemas read lock: {}",
-                e
-            );
-            return HttpResponse::InternalServerError().json(ErrorResponse {
-                error: "Failed to acquire schemas read lock".to_string(),
-            });
-        }
-    };
-
-    let schema_list: Vec<Schema> = schemas.values().cloned().collect();
-
-    HttpResponse::Ok().json(AvailableSchemasResponse {
-        schemas: schema_list,
-    })
-}
-
-/// Get a specific schema by name
-async fn get_schema(
-    path: web::Path<String>,
-    state: web::Data<SchemaServiceState>,
-) -> impl Responder {
-    let schema_name = path.into_inner();
-    log_feature!(
-        LogFeature::Schema,
-        info,
-        "Schema service: getting schema '{}'",
-        schema_name
-    );
-
-    let schemas = match state.schemas.read() {
-        Ok(s) => s,
-        Err(e) => {
-            log_feature!(
-                LogFeature::Schema,
-                error,
-                "Failed to acquire schemas read lock: {}",
-                e
-            );
-            return HttpResponse::InternalServerError().json(ErrorResponse {
-                error: "Failed to acquire schemas read lock".to_string(),
-            });
-        }
-    };
-
-    match schemas.get(&schema_name) {
-        Some(schema) => HttpResponse::Ok().json(schema),
-        None => {
-            log_feature!(
-                LogFeature::Schema,
-                warn,
-                "Schema '{}' not found",
-                schema_name
-            );
-            HttpResponse::NotFound().json(ErrorResponse {
-                error: "Schema not found".to_string(),
-            })
-        }
-    }
-}
-
-/// Query parameters for the find-similar endpoint
-#[derive(Debug, Deserialize)]
-struct SimilarQuery {
-    threshold: Option<f64>,
-}
-
-/// Find schemas similar to the given schema
-async fn find_similar(
-    path: web::Path<String>,
-    query: web::Query<SimilarQuery>,
-    state: web::Data<SchemaServiceState>,
-) -> impl Responder {
-    let schema_name = path.into_inner();
-    let threshold = query.threshold.unwrap_or(0.5);
-
-    if !(0.0..=1.0).contains(&threshold) {
-        return HttpResponse::BadRequest().json(ErrorResponse {
-            error: "Threshold must be between 0.0 and 1.0".to_string(),
-        });
-    }
-
-    log_feature!(
-        LogFeature::Schema,
-        info,
-        "Schema service: finding schemas similar to '{}' with threshold {}",
-        schema_name,
-        threshold
-    );
-
-    match state.find_similar_schemas(&schema_name, threshold) {
-        Ok(response) => HttpResponse::Ok().json(response),
-        Err(e) => {
-            let error_msg = format!("{}", e);
-            if error_msg.contains("not found") {
-                HttpResponse::NotFound().json(ErrorResponse {
-                    error: format!("Schema '{}' not found", schema_name),
-                })
-            } else {
-                HttpResponse::InternalServerError().json(ErrorResponse {
-                    error: format!("Failed to find similar schemas: {}", e),
-                })
-            }
-        }
-    }
-}
-
-/// Reload schemas from the directory
-async fn reload_schemas(state: web::Data<SchemaServiceState>) -> impl Responder {
-    log_feature!(
-        LogFeature::Schema,
-        info,
-        "Schema service: reloading schemas"
-    );
-
-    match state.load_schemas().await {
-        Ok(_) => {
-            let schemas = match state.schemas.read() {
-                Ok(s) => s,
-                Err(e) => {
-                    log_feature!(
-                        LogFeature::Schema,
-                        error,
-                        "Failed to acquire schemas read lock: {}",
-                        e
-                    );
-                    return HttpResponse::InternalServerError().json(ErrorResponse {
-                        error: "Failed to acquire schemas read lock".to_string(),
-                    });
-                }
-            };
-
-            HttpResponse::Ok().json(ReloadResponse {
-                success: true,
-                schemas_loaded: schemas.len(),
-            })
-        }
-        Err(e) => {
-            log_feature!(LogFeature::Schema, error, "Failed to reload schemas: {}", e);
-            HttpResponse::InternalServerError().json(ErrorResponse {
-                error: format!("Failed to reload schemas: {}", e),
-            })
-        }
-    }
-}
-
-async fn add_schema(
-    payload: web::Json<AddSchemaRequest>,
-    state: web::Data<SchemaServiceState>,
-) -> impl Responder {
-    let request = payload.into_inner();
-    let schema_name = request.schema.name.clone();
-
-    log_feature!(
-        LogFeature::Schema,
-        info,
-        "Schema service: adding schema '{}' with {} mutation mappers",
-        schema_name,
-        request.mutation_mappers.len()
-    );
-
-    match state
-        .add_schema(request.schema, request.mutation_mappers)
-        .await
-    {
-        Ok(SchemaAddOutcome::Added(schema, mutation_mappers)) => {
-            HttpResponse::Created().json(AddSchemaResponse {
-                schema,
-                mutation_mappers,
-            })
-        }
-        Ok(SchemaAddOutcome::AlreadyExists(schema)) => {
-            HttpResponse::Ok().json(AddSchemaResponse {
-                schema,
-                mutation_mappers: HashMap::new(),
-            })
-        }
-        Ok(SchemaAddOutcome::TooSimilar(conflict)) => {
-            HttpResponse::Conflict().json(ConflictResponse {
-                error: "Schema too similar to existing schema".to_string(),
-                similarity: conflict.similarity,
-                closest_schema: conflict.closest_schema,
-            })
-        }
-        Err(error) => {
-            log_feature!(
-                LogFeature::Schema,
-                error,
-                "Failed to add schema '{}': {}",
-                schema_name,
-                error
-            );
-            HttpResponse::BadRequest().json(ErrorResponse {
-                error: format!("Failed to add schema: {}", error),
-            })
-        }
-    }
-}
-
-/// Health check endpoint
-async fn health_check() -> impl Responder {
-    HttpResponse::Ok().json(HealthResponse {
-        status: "healthy".to_string(),
-    })
-}
-
-/// Request for resetting the schema service database
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ResetRequest {
-    pub confirm: bool,
-}
-
-/// Response for resetting the schema service database
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ResetResponse {
-    pub success: bool,
-    pub message: String,
-}
-
-/// Reset the schema service database
-async fn reset_database(
-    state: web::Data<SchemaServiceState>,
-    req: web::Json<ResetRequest>,
-) -> impl Responder {
-    // Require explicit confirmation
-    if !req.confirm {
-        return HttpResponse::BadRequest().json(ResetResponse {
-            success: false,
-            message: "Reset confirmation required. Set 'confirm' to true.".to_string(),
-        });
-    }
-
-    log_feature!(
-        LogFeature::Schema,
-        info,
-        "Resetting schema service database"
-    );
-
-    // Clear the in-memory schemas map
-    {
-        let mut schemas = match state.schemas.write() {
-            Ok(s) => s,
-            Err(e) => {
-                log_feature!(
-                    LogFeature::Schema,
-                    error,
-                    "Failed to acquire schemas write lock: {}",
-                    e
-                );
-                return HttpResponse::InternalServerError().json(ResetResponse {
-                    success: false,
-                    message: "Failed to acquire schemas write lock".to_string(),
-                });
-            }
-        };
-        schemas.clear();
-    }
-
-    // Clear storage backend
-    match &state.storage {
-        SchemaStorage::Sled { db, schemas_tree } => {
-            // Clear all entries from the schemas tree
-            if let Err(e) = schemas_tree.clear() {
-                log_feature!(
-                    LogFeature::Schema,
-                    error,
-                    "Failed to clear schemas tree: {}",
-                    e
-                );
-                return HttpResponse::InternalServerError().json(ResetResponse {
-                    success: false,
-                    message: format!("Failed to reset sled database: {}", e),
-                });
-            }
-
-            // Flush to ensure changes are persisted
-            if let Err(e) = db.flush() {
-                log_feature!(
-                    LogFeature::Schema,
-                    warn,
-                    "Failed to flush database after reset: {}",
-                    e
-                );
-            }
-        }
-        #[cfg(feature = "aws-backend")]
-        SchemaStorage::Cloud { store } => {
-            // Clear all schemas from DynamoDB
-            if let Err(e) = store.clear_all_schemas().await {
-                log_feature!(
-                    LogFeature::Schema,
-                    error,
-                    "Failed to clear DynamoDB schemas: {}",
-                    e
-                );
-                return HttpResponse::InternalServerError().json(ResetResponse {
-                    success: false,
-                    message: format!("Failed to reset DynamoDB: {}", e),
-                });
-            }
-        }
-    }
-
-    log_feature!(
-        LogFeature::Schema,
-        info,
-        "Schema service database reset successfully"
-    );
-
-    HttpResponse::Ok().json(ResetResponse {
-        success: true,
-        message: "Schema service database reset successfully. All schemas have been cleared."
-            .to_string(),
-    })
-}
+pub use super::state::CloudConfig;
+
+// Route handlers (pub(super) visibility — accessible from sibling modules)
+use super::routes::{
+    add_schema, find_similar, get_available_schemas, get_schema, health_check, list_schemas,
+    reload_schemas, reset_database,
+};
 
 /// Schema Service HTTP Server
 pub struct SchemaServiceServer {
@@ -939,6 +94,9 @@ impl SchemaServiceServer {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::schema::types::Schema;
+    use crate::schema_service::state::jaccard_index;
+    use std::collections::{HashMap, HashSet};
     use tempfile::tempdir;
 
     #[tokio::test]
@@ -994,7 +152,6 @@ mod tests {
         // Check stored by combined name
         assert!(stored_schemas.contains_key(&added_schema.name));
 
-        // Check the underlying storage
         // Check the underlying storage
         match &state.storage {
             SchemaStorage::Sled { schemas_tree, .. } => {
