+++ src/schema_service/state.rs
@@ -1,0 +1,435 @@
+use std::collections::{HashMap, HashSet};
+use std::sync::{Arc, RwLock};
+
+use crate::error::{FoldDbError, FoldDbResult};
+use crate::log_feature;
+use crate::logging::features::LogFeature;
+use crate::schema::types::Schema;
+#[cfg(feature = "aws-backend")]
+use crate::storage::DynamoDbSchemaStore;
+
+#[cfg(feature = "aws-backend")]
+pub use crate::storage::CloudConfig;
+
+use super::types::{SchemaAddOutcome, SimilarSchemaEntry, SimilarSchemasResponse};
+
+/// Storage backend for the schema service
+#[derive(Clone)]
+pub enum SchemaStorage {
+    /// Local sled database (default)
+    Sled {
+        db: sled::Db,
+        schemas_tree: sled::Tree,
+    },
+    /// Cloud storage (DynamoDB etc) (serverless, no locking needed!)
+    #[cfg(feature = "aws-backend")]
+    Cloud { store: Arc<DynamoDbSchemaStore> },
+}
+
+/// Shared state for the schema service
+#[derive(Clone)]
+pub struct SchemaServiceState {
+    pub(super) schemas: Arc<RwLock<HashMap<String, Schema>>>,
+    pub(super) storage: SchemaStorage,
+}
+
+impl SchemaServiceState {
+    /// Create a new schema service state with local sled storage
+    pub fn new(db_path: String) -> FoldDbResult<Self> {
+        let db = sled::open(&db_path).map_err(|e| {
+            FoldDbError::Config(format!(
+                "Failed to open schema service database at '{}': {}",
+                db_path, e
+            ))
+        })?;
+
+        let schemas_tree = db
+            .open_tree("schemas")
+            .map_err(|e| FoldDbError::Config(format!("Failed to open schemas tree: {}", e)))?;
+
+        let state = Self {
+            schemas: Arc::new(RwLock::new(HashMap::new())),
+            storage: SchemaStorage::Sled { db, schemas_tree },
+        };
+
+        // Load schemas synchronously for sled
+        state.load_schemas_sync()?;
+
+        Ok(state)
+    }
+
+    /// Synchronous version of load_schemas for Sled storage
+    fn load_schemas_sync(&self) -> FoldDbResult<()> {
+        let mut schemas = self
+            .schemas
+            .write()
+            .map_err(|_| FoldDbError::Config("Failed to acquire schemas write lock".to_string()))?;
+
+        schemas.clear();
+
+        match &self.storage {
+            SchemaStorage::Sled { schemas_tree, .. } => {
+                let mut count = 0;
+                for result in schemas_tree.iter() {
+                    let (key, value) = result.map_err(|e| {
+                        FoldDbError::Config(format!("Failed to iterate over schemas tree: {}", e))
+                    })?;
+
+                    let name = String::from_utf8(key.to_vec()).map_err(|e| {
+                        FoldDbError::Config(format!("Failed to decode schema name from key: {}", e))
+                    })?;
+
+                    let schema: Schema = serde_json::from_slice(&value).map_err(|e| {
+                        FoldDbError::Config(format!(
+                            "Failed to parse schema '{}' from database: {}",
+                            name, e
+                        ))
+                    })?;
+
+                    schemas.insert(name, schema);
+                    count += 1;
+                }
+
+                log_feature!(
+                    LogFeature::Schema,
+                    info,
+                    "Schema service loaded {} schemas from sled",
+                    count
+                );
+            }
+            #[cfg(feature = "aws-backend")]
+            _ => {
+                return Err(FoldDbError::Config(
+                    "load_schemas_sync called on non-Sled storage".to_string(),
+                ));
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Create a new schema service state with Cloud storage
+    /// No locking needed - identity hashes ensure idempotent writes!
+    #[cfg(feature = "aws-backend")]
+    pub async fn new_with_cloud(config: CloudConfig) -> FoldDbResult<Self> {
+        log_feature!(
+            LogFeature::Schema,
+            info,
+            "Initializing schema service with DynamoDB in region: {}",
+            config.region
+        );
+
+        let store = DynamoDbSchemaStore::new(config).await?;
+
+        let state = Self {
+            schemas: Arc::new(RwLock::new(HashMap::new())),
+            storage: SchemaStorage::Cloud {
+                store: Arc::new(store),
+            },
+        };
+
+        // Load schemas on initialization
+        state.load_schemas().await?;
+
+        log_feature!(
+            LogFeature::Schema,
+            info,
+            "Schema service initialized with DynamoDB, loaded {} schemas",
+            state.schemas.read().map(|s| s.len()).unwrap_or(0)
+        );
+
+        Ok(state)
+    }
+
+    /// Load all schemas from storage (works for both Sled and DynamoDB)
+    pub async fn load_schemas(&self) -> FoldDbResult<()> {
+        match &self.storage {
+            SchemaStorage::Sled { schemas_tree, .. } => {
+                let mut schemas = self.schemas.write().map_err(|_| {
+                    FoldDbError::Config("Failed to acquire schemas write lock".to_string())
+                })?;
+
+                schemas.clear();
+                let mut count = 0;
+                for result in schemas_tree.iter() {
+                    let (key, value) = result.map_err(|e| {
+                        FoldDbError::Config(format!("Failed to iterate over schemas tree: {}", e))
+                    })?;
+
+                    let name = String::from_utf8(key.to_vec()).map_err(|e| {
+                        FoldDbError::Config(format!("Failed to decode schema name from key: {}", e))
+                    })?;
+
+                    let schema: Schema = serde_json::from_slice(&value).map_err(|e| {
+                        FoldDbError::Config(format!(
+                            "Failed to parse schema '{}' from database: {}",
+                            name, e
+                        ))
+                    })?;
+
+                    log_feature!(
+                        LogFeature::Schema,
+                        info,
+                        "Loaded schema '{}' from sled database",
+                        name
+                    );
+
+                    schemas.insert(name, schema);
+                    count += 1;
+                }
+
+                log_feature!(
+                    LogFeature::Schema,
+                    info,
+                    "Schema service loaded {} schemas from sled",
+                    count
+                );
+            }
+            #[cfg(feature = "aws-backend")]
+            SchemaStorage::Cloud { store } => {
+                let all_schemas = store.get_all_schemas().await?;
+                let count = all_schemas.len();
+
+                let mut schemas = self.schemas.write().map_err(|_| {
+                    FoldDbError::Config("Failed to acquire schemas write lock".to_string())
+                })?;
+
+                schemas.clear();
+
+                for schema in all_schemas {
+                    log_feature!(
+                        LogFeature::Schema,
+                        info,
+                        "Loaded schema '{}' from DynamoDB",
+                        schema.name
+                    );
+                    schemas.insert(schema.name.clone(), schema);
+                }
+
+                log_feature!(
+                    LogFeature::Schema,
+                    info,
+                    "Schema service loaded {} schemas from DynamoDB",
+                    count
+                );
+            }
+        }
+
+        Ok(())
+    }
+
+    pub async fn add_schema(
+        &self,
+        mut schema: Schema,
+        mutation_mappers: HashMap<String, String>,
+    ) -> FoldDbResult<SchemaAddOutcome> {
+        // Ensure identity_hash is computed
+        if schema.identity_hash.is_none() {
+            schema.compute_identity_hash();
+        }
+
+        // Get the original schema name before we modify it
+        let original_schema_name = schema.name.clone();
+
+        // Use identity_hash as the schema identifier
+        let identity_hash = schema
+            .get_identity_hash()
+            .ok_or_else(|| {
+                FoldDbError::Config("Schema must have identity_hash computed".to_string())
+            })?
+            .clone();
+
+        log_feature!(
+            LogFeature::Schema,
+            info,
+            "Schema '{}' identity_hash: {}",
+            original_schema_name,
+            identity_hash
+        );
+
+        // Use identity_hash as unique identifier (SHA256 of sorted field names)
+        let schema_name = identity_hash.clone();
+
+        // Check if this exact schema already exists (same identity hash = same fields)
+        {
+            let schemas = self.schemas.read().map_err(|_| {
+                FoldDbError::Config("Failed to acquire schemas read lock".to_string())
+            })?;
+
+            if let Some(existing_schema) = schemas.get(&schema_name) {
+                log_feature!(
+                    LogFeature::Schema,
+                    info,
+                    "Schema '{}' already exists - returning existing schema",
+                    schema_name
+                );
+
+                return Ok(SchemaAddOutcome::AlreadyExists(existing_schema.clone()));
+            }
+        }
+
+        schema.name = schema_name.clone();
+
+        // Persist to storage backend
+        match &self.storage {
+            SchemaStorage::Sled { db, schemas_tree } => {
+                let serialized_schema = serde_json::to_vec(&schema).map_err(|error| {
+                    FoldDbError::Serialization(format!(
+                        "Failed to serialize schema '{}': {}",
+                        schema_name, error
+                    ))
+                })?;
+
+                schemas_tree
+                    .insert(schema_name.as_bytes(), serialized_schema)
+                    .map_err(|error| {
+                        FoldDbError::Config(format!(
+                            "Failed to insert schema '{}' into sled database: {}",
+                            schema_name, error
+                        ))
+                    })?;
+
+                db.flush().map_err(|error| {
+                    FoldDbError::Config(format!("Failed to flush sled database: {}", error))
+                })?;
+
+                log_feature!(
+                    LogFeature::Schema,
+                    info,
+                    "Schema '{}' persisted to sled database",
+                    schema_name
+                );
+            }
+            #[cfg(feature = "aws-backend")]
+            SchemaStorage::Cloud { store } => {
+                // No locking needed! Identity hash ensures idempotent writes
+                store.put_schema(&schema, &mutation_mappers).await?;
+
+                log_feature!(
+                    LogFeature::Schema,
+                    info,
+                    "Schema '{}' persisted to DynamoDB (no locking needed!)",
+                    schema_name
+                );
+            }
+        }
+
+        // Insert into in-memory cache
+        {
+            let mut schemas = self.schemas.write().map_err(|_| {
+                FoldDbError::Config("Failed to acquire schemas write lock".to_string())
+            })?;
+            schemas.insert(schema_name.clone(), schema.clone());
+        }
+
+        log_feature!(
+            LogFeature::Schema,
+            info,
+            "Schema '{}' successfully added to registry",
+            schema_name
+        );
+
+        Ok(SchemaAddOutcome::Added(schema, mutation_mappers))
+    }
+
+    /// Get all schema names (public accessor for Lambda integration)
+    pub fn get_schema_names(&self) -> FoldDbResult<Vec<String>> {
+        let schemas = self
+            .schemas
+            .read()
+            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
+        Ok(schemas.keys().cloned().collect())
+    }
+
+    /// Get all schemas (public accessor for Lambda integration)
+    pub fn get_all_schemas_cached(&self) -> FoldDbResult<Vec<Schema>> {
+        let schemas = self
+            .schemas
+            .read()
+            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
+        Ok(schemas.values().cloned().collect())
+    }
+
+    /// Get a schema by name (public accessor for Lambda integration)
+    pub fn get_schema_by_name(&self, name: &str) -> FoldDbResult<Option<Schema>> {
+        let schemas = self
+            .schemas
+            .read()
+            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
+        Ok(schemas.get(name).cloned())
+    }
+
+    /// Get schema count (public accessor for Lambda integration)
+    pub fn get_schema_count(&self) -> usize {
+        self.schemas.read().map(|s| s.len()).unwrap_or(0)
+    }
+
+    /// Find schemas similar to the given schema using Jaccard index on field name sets
+    pub fn find_similar_schemas(
+        &self,
+        name: &str,
+        threshold: f64,
+    ) -> FoldDbResult<SimilarSchemasResponse> {
+        let schemas = self
+            .schemas
+            .read()
+            .map_err(|_| FoldDbError::Config("Failed to acquire schemas read lock".to_string()))?;
+
+        let target = schemas.get(name).ok_or_else(|| {
+            FoldDbError::Config(format!("Schema '{}' not found", name))
+        })?;
+
+        let target_fields = collect_field_names(target);
+
+        let mut similar: Vec<SimilarSchemaEntry> = schemas
+            .iter()
+            .filter(|(k, _)| k.as_str() != name)
+            .filter_map(|(_, schema)| {
+                let other_fields = collect_field_names(schema);
+                let similarity = jaccard_index(&target_fields, &other_fields);
+                if similarity >= threshold {
+                    Some(SimilarSchemaEntry {
+                        schema: schema.clone(),
+                        similarity,
+                    })
+                } else {
+                    None
+                }
+            })
+            .collect();
+
+        similar.sort_by(|a, b| b.similarity.partial_cmp(&a.similarity).unwrap_or(std::cmp::Ordering::Equal));
+
+        Ok(SimilarSchemasResponse {
+            query_schema: name.to_string(),
+            threshold,
+            similar_schemas: similar,
+        })
+    }
+}
+
+/// Collect all field names from a schema (union of fields and transform_fields keys)
+fn collect_field_names(schema: &Schema) -> HashSet<String> {
+    let mut names = HashSet::new();
+    if let Some(ref fields) = schema.fields {
+        for f in fields {
+            names.insert(f.clone());
+        }
+    }
+    if let Some(ref tf) = schema.transform_fields {
+        for key in tf.keys() {
+            names.insert(key.clone());
+        }
+    }
+    names
+}
+
+/// Compute Jaccard index: |A ∩ B| / |A ∪ B|
+pub(crate) fn jaccard_index(a: &HashSet<String>, b: &HashSet<String>) -> f64 {
+    if a.is_empty() && b.is_empty() {
+        return 1.0;
+    }
+    let intersection = a.intersection(b).count();
+    let union = a.union(b).count();
+    intersection as f64 / union as f64
+}
